requirements.txt
setuptools~=50.3.0
PyInquirer
python-dateutil